+++ src/main/java/duke/Storage.java
@@ -1,10 +1,17 @@
+package duke;
 import java.io.*;
-import java.util.ArrayList;
+import duke.TaskList;
 
-public class Save {
-    public static void saveTaskList(TaskList tl) {
+public class Storage {
+    private String filePath;
+
+    public Storage(String filePath) {
+        this.filePath = filePath;
+    }
+
+    public void saveTaskList(TaskList tl) {
         try {
-            FileOutputStream fos = new FileOutputStream("taskList.txt");
+            FileOutputStream fos = new FileOutputStream(this.filePath);
             ObjectOutputStream oos = new ObjectOutputStream(fos);
             oos.writeObject(tl);
             oos.close();
@@ -15,12 +22,12 @@ public class Save {
         }
     }
 
-    public static TaskList loadTaskList() {
-        File file = new File("taskList.txt");
+    public TaskList loadTaskList() {
+        File file = new File(this.filePath);
         if (file.exists()) {
             TaskList tl = new TaskList();
             try {
-                FileInputStream fis = new FileInputStream("taskList.txt");
+                FileInputStream fis = new FileInputStream(this.filePath);
                 ObjectInputStream ois = new ObjectInputStream(fis);
                 tl = (TaskList) ois.readObject();
                 ois.close();
