+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,5 @@
+package duke;
+import duke.tasks.Task;
 import java.io.Serializable;
 import java.util.ArrayList;
