+++ src/main/java/duke/tasks/Deadline.java
@@ -1,10 +1,11 @@
+package duke.tasks;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
-import java.util.Date;
 
-public class Deadline extends Task{
+public class Deadline extends Task {
     protected String by;
 
     public Deadline(String description, String date) {
