+++ src/main/java/duke/tasks/Event.java
@@ -1,3 +1,5 @@
+package duke.tasks;
+
 import java.io.Serializable;
 
 public class Event extends Task {
